Move time provider util to utils folder
